# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/listings": {
      "get": {
        "operationId": "PublicListingsController_cards_v1",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "description": "Up to 20 listing ids, comma-separated",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingCardDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Cards for up to 20 live listings, in the order given (recently viewed)",
        "tags": [
          "listings"
        ]
      }
    },
    "/v1/listings/{id}": {
      "get": {
        "operationId": "PublicListingsController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicListingDto"
                }
              }
            }
          }
        },
        "summary": "A live listing (public; approximate location only). Counts a view for “Popular”.",
        "tags": [
          "listings"
        ]
      }
    },
    "/v1/listings/{id}/quote": {
      "get": {
        "operationId": "PublicListingsController_quote_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-10-02",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-10-05",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteDto"
                }
              }
            }
          }
        },
        "summary": "Price breakdown for dates, and whether the item is free then",
        "tags": [
          "listings"
        ]
      }
    },
    "/v1/me/listings": {
      "get": {
        "operationId": "MyListingsController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyListingDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me · listings"
        ]
      },
      "post": {
        "operationId": "MyListingsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          },
          "403": {
            "description": "VERIFICATION_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a draft listing (needs a verified phone and email)",
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}": {
      "get": {
        "operationId": "MyListingsController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me · listings"
        ]
      },
      "patch": {
        "operationId": "MyListingsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit; a rejected listing goes back to draft",
        "tags": [
          "me · listings"
        ]
      },
      "delete": {
        "operationId": "MyListingsController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a listing and its photos",
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}/photos": {
      "post": {
        "operationId": "MyListingsController_addPhoto_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          },
          "409": {
            "description": "LISTING_PHOTO_LIMIT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a photo from an upload (purpose LISTING_PHOTO); max 8",
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}/photos/order": {
      "put": {
        "operationId": "MyListingsController_reorderPhotos_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reorder photos (cover first)",
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}/photos/{photoId}": {
      "delete": {
        "operationId": "MyListingsController_deletePhoto_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}/blocks": {
      "put": {
        "operationId": "MyListingsController_setBlocks_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlocksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the blocked dates (inclusive ranges, next 365 days)",
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}/required-docs": {
      "put": {
        "operationId": "MyListingsController_setRequiredDocs_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequiredDocsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the documents a borrower must share",
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}/publish": {
      "post": {
        "operationId": "MyListingsController_publish_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResultDto"
                }
              }
            }
          },
          "400": {
            "description": "LISTING_INCOMPLETE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a draft: your first listing is reviewed; later ones go live at once",
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}/pause": {
      "post": {
        "operationId": "MyListingsController_pause_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/me/listings/{id}/unpause": {
      "post": {
        "operationId": "MyListingsController_unpause_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me · listings"
        ]
      }
    },
    "/v1/admin/listings": {
      "get": {
        "operationId": "AdminListingsController_list_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "PENDING",
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING",
                "LIVE",
                "PAUSED",
                "REJECTED",
                "REMOVED",
                "DELETED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Title, or the lender’s phone or name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListingPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listings by status (pending: oldest first), with search",
        "tags": [
          "admin · listings"
        ]
      }
    },
    "/v1/admin/listings/{id}": {
      "get": {
        "operationId": "AdminListingsController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · listings"
        ]
      }
    },
    "/v1/admin/listings/{id}/approve": {
      "post": {
        "operationId": "AdminListingsController_approve_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListingDto"
                }
              }
            }
          },
          "409": {
            "description": "LISTING_STATUS_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · listings"
        ]
      }
    },
    "/v1/admin/listings/{id}/reject": {
      "post": {
        "operationId": "AdminListingsController_reject_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListingDto"
                }
              }
            }
          },
          "409": {
            "description": "LISTING_STATUS_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · listings"
        ]
      }
    },
...
    "/v1/admin/listings/{id}/category": {
      "patch": {
        "operationId": "AdminListingsController_changeCategory_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · listings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sajha API",
    "description": "REST API for the Sajha mobile app, admin panel and website.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "QuoteDto": {
        "type": "object",
        "properties": {
          "days": {
            "type": "number",
            "description": "Inclusive: pickup day to return day"
          },
          "pricePerDayPaise": {
            "type": "number"
          },
          "rentBeforeDiscountPaise": {
            "type": "number"
          },
          "weeklyDiscountPaise": {
            "type": "number"
          },
          "rentPaise": {
            "type": "number"
          },
          "feePaise": {
            "type": "number",
            "description": "Borrower service fee (₹0 for now)"
          },
          "depositPaise": {
            "type": "number",
            "description": "Refundable"
          },
          "totalPaise": {
            "type": "number"
          },
          "available": {
            "type": "boolean"
          },
          "unavailableReason": {
            "type": "string",
            "enum": [
              "BLOCKED",
              "TOO_SHORT",
              "TOO_LONG",
              "NOT_ENOUGH_NOTICE"
            ],
            "nullable": true
          }
        },
        "required": [
          "days",
          "pricePerDayPaise",
          "rentBeforeDiscountPaise",
          "weeklyDiscountPaise",
          "rentPaise",
          "feePaise",
          "depositPaise",
          "totalPaise",
          "available"
        ]
      },
...
    }
  }
}